Moved functions in/out of common areas
