+++ yk_gmd_blender/blender/importer/scene_creators/base.py
@@ -4,17 +4,21 @@ from dataclasses import dataclass
 from enum import Enum
 from typing import Dict, Union
 
+import bpy
 import bmesh
 from mathutils import Vector, Matrix
 
-import bpy
-from yk_gmd_blender.blender.common import root_name_for_gmd_file, GMDGame
+from yk_gmd_blender.blender.common import GMDGame
 from yk_gmd_blender.blender.importer.mesh_importer import gmd_meshes_to_bmesh
 from yk_gmd_blender.blender.materials import get_yakuza_shader_node_group, set_yakuza_shader_material_from_attributeset
 from yk_gmd_blender.yk_gmd.v2.abstract.gmd_attributes import GMDAttributeSet
 from yk_gmd_blender.yk_gmd.v2.abstract.gmd_scene import GMDScene
 from yk_gmd_blender.yk_gmd.v2.abstract.nodes.gmd_object import GMDSkinnedObject, GMDUnskinnedObject
 from yk_gmd_blender.yk_gmd.v2.errors.error_reporter import ErrorReporter
+
+
+def root_name_for_gmd_file(gmd_file: GMDScene):
+    return f"{gmd_file.name}"
 
 
 class MaterialNamingType(Enum):
