+++ yk_gmd_blender/blender/importer/scene_creators/unskinned.py
@@ -1,8 +1,8 @@
 from mathutils import Quaternion
 
 import bpy
-from yk_gmd_blender.blender.common import root_name_for_gmd_file
-from yk_gmd_blender.blender.importer.scene_creators.base import BaseGMDSceneCreator, GMDSceneCreatorConfig
+from yk_gmd_blender.blender.importer.scene_creators.base import BaseGMDSceneCreator, GMDSceneCreatorConfig, \
+    root_name_for_gmd_file
 from yk_gmd_blender.yk_gmd.v2.abstract.gmd_scene import GMDScene
 from yk_gmd_blender.yk_gmd.v2.abstract.nodes.gmd_bone import GMDBone
 from yk_gmd_blender.yk_gmd.v2.abstract.nodes.gmd_object import GMDSkinnedObject, GMDUnskinnedObject
